apply targeting changes inline
