+++ packages/back-end/src/services/holdouts.ts
@@ -13,7 +13,6 @@ import {
   CreateHoldoutInput,
   HoldoutInterface,
   HoldoutNextScheduledStatusUpdate,
-  SavedGroupTargeting,
 } from "shared/validators";
 import {
   Changeset,
@@ -45,7 +44,6 @@ import {
   validateExperimentData,
   validateVariationIds,
 } from "back-end/src/services/experiments";
-import { validateExperimentChange } from "./experimentChanges/changeExperimentStatus";
 
 export async function canLinkExperimentToHoldoutFromFeatures(
   context: ReqContext | ApiReqContext,
@@ -560,51 +558,6 @@ export async function setHoldoutStage(
       stage satisfies never;
     }
   }
-}
-
-/**
- * Applies phase-level targeting and sizing changes to a holdout's companion
- * experiment through the same validate/update path the internal targeting
- * endpoint uses, so the stored phases and the SDK payload stay correct.
- * Only the fields that are provided are changed. Callers are responsible for
- * the `canRunHoldout` permission check.
- */
-export async function applyHoldoutTargetingChanges(
-  context: ReqContext | ApiReqContext,
-  {
-    experiment,
-    coverage,
-    condition,
-    savedGroups,
-    hashAttribute,
-  }: {
-    experiment: ExperimentInterface;
-    coverage?: number;
-    condition?: string;
-    savedGroups?: SavedGroupTargeting[];
-    hashAttribute?: string;
-  },
-): Promise<ExperimentInterface> {
-  const phases = [...experiment.phases];
-  if (!phases.length) {
-    throw new Error("Holdout does not have a phase to target");
-  }
-
-  const current = phases[phases.length - 1];
-  phases[phases.length - 1] = {
-    ...current,
-    condition: condition ?? current.condition,
-    savedGroups: savedGroups ?? current.savedGroups,
-    coverage: coverage ?? current.coverage,
-  };
-
-  const changes: Changeset = { phases };
-  if (hashAttribute !== undefined) {
-    changes.hashAttribute = hashAttribute;
-  }
-
-  await validateExperimentChange({ context, experiment, changes });
-  return updateExperiment({ context, experiment, changes });
 }
 
 /**
